Moved AccountDataSynchronizationService from application layer to presentation layer
