--- code/FinanceManager.Application/Services/AccountDataSynchronizationService.cs
@@ -1,12 +1,0 @@
-﻿namespace FinanceManager.Application.Services
-{
-    public class AccountDataSynchronizationService
-    {
-        public event Action? AccountsChanged;
-        public async Task AccountChanged()
-        {
-            AccountsChanged?.Invoke();
-            await Task.CompletedTask;
-        }
-    }
-}
