+++ code/FinanceManager.Components/Components/AccountDetailsPageContents/BankAccountComponents/BankAccountDetailsPageContent.razor.cs
@@ -1,6 +1,6 @@
 using ApexCharts;
-using FinanceManager.Application.Services;
 using FinanceManager.Components.Helpers;
+using FinanceManager.Components.Services;
 using FinanceManager.Domain.Entities.Accounts;
 using FinanceManager.Domain.Entities.Accounts.Entries;
 using FinanceManager.Domain.Entities.Login;
