+++ code/FinanceManager.Components/Components/AccountDetailsPageContents/BankAccountComponents/ManageBankAccount.razor.cs
@@ -1,4 +1,5 @@
-using FinanceManager.Application.Services;
+
+using FinanceManager.Components.Services;
 using FinanceManager.Domain.Entities.Accounts;
 using FinanceManager.Domain.Enums;
 using FinanceManager.Domain.Repositories.Account;
