+++ code/FinanceManager.Components/Components/AccountDetailsPageContents/StockAccountComponents/StockAccountDetailsPageContent.razor.cs
@@ -1,6 +1,6 @@
 using ApexCharts;
-using FinanceManager.Application.Services;
 using FinanceManager.Components.Helpers;
+using FinanceManager.Components.Services;
 using FinanceManager.Domain.Entities;
 using FinanceManager.Domain.Entities.Accounts;
 using FinanceManager.Domain.Entities.Accounts.Entries;
