+++ code/FinanceManager.Components/Services/AccountDataSynchronizationService.cs
@@ -1,0 +1,11 @@
+﻿namespace FinanceManager.Components.Services;
+
+public class AccountDataSynchronizationService
+{
+    public event Action? AccountsChanged;
+    public async Task AccountChanged()
+    {
+        AccountsChanged?.Invoke();
+        await Task.CompletedTask;
+    }
+}
