+++ code/FinanceManager.Components/_Imports.razor
@@ -2,3 +2,5 @@
 @using FinanceManager.Domain.Services
 @using MudBlazor
 @using Microsoft.AspNetCore.Components.Authorization
+
+@using FinanceManager.Components.Services;
